requirements.txt
pyautogui>=0.9.50
Pillow>=8.0
